change utils folder structure
