+++ src/utils/getCurrentSlideIndex.ts
@@ -1,0 +1,13 @@
+import { CarouselConfig } from '../types'
+
+export function getCurrentSlideIndex(
+  config: CarouselConfig,
+  val: number,
+  max: number,
+  min: number
+): number {
+  if (config.wrapAround) {
+    return val
+  }
+  return Math.min(Math.max(val, min), max)
+}
