+++ src/utils/index.ts
@@ -1,0 +1,8 @@
+export * from './debounce'
+export * from './getCurrentSlideIndex'
+export * from './getMaxSlideIndex'
+export * from './getMinSlideIndex'
+export * from './getSlidesToScroll'
+export * from './getSlidesVNodes'
+export * from './mapNumberToRange'
+export * from './throttle'
